Stack02
here we are check is the given string a palindrome or not . but here we have time complexity problem. 
Just try to understand the code and run it. In next file (Stack 03) we have a good example how can we get rid of the problem.
in Stack 03 we have less time complexity which is log(n/2).
read the code and run. If you face any problem just do comment below.
Will try to get back to you soon.
Thanks